6.19改后：最终版本 改动见文档

+++ Classes/Entity/Weapon/Ammo/Ammo.h
@@ -6,7 +6,7 @@ class Ammo : public Entity
 {
 public:
 
-	double start;
+	long start;
 
 private:
